PDFBOX-3991: deprecate methods with double parameters, add methods with float parameters
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/trunk@1814321 13f79535-47bb-0310-9956-ffa450edef68

+++ pdfbox/src/main/java/org/apache/pdfbox/pdmodel/PDPageContentStream.java
@@ -1115,7 +1115,7 @@ public final class PDPageContentStream extends PDAbstractContentStream implement
     @Deprecated
     public void appendRawCommands(double data) throws IOException
     {
-        writeOperand(data);
+        writeOperand((float) data);
     }
 
     /**
